Refactor weather DTOs and add OccurencyQuery class
- Minor formatting change in `WeatherForecastDto` for `Latitude`.
- Removed `WeatherSliceDto` from `WeatherForecastDto.cs`.
- Updated property casing for `rainVolumeMm` in `OpenWeatherClient.cs`.
- Added new `OccurencyQuery` class for filtering activity occurrences.
- Introduced `WeatherSliceDto` class with weather data properties.
