+++ Backend/Api/Application/ActivityOccurrence/DTO/Request/OccurencyQuery.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Application.ActivityOccurrence.DTO.Request
+{
+    public sealed class OccurencyQuery
+    {
+        // Filter parameters
+        // Filter by date range
+        public DateTime FromDate { get; init; }
+        public DateTime ToDate { get; init; }
+
+        // Optional filters
+        public Guid? CategoryId { get; init; }
+        public Guid? ActivityId { get; init; }   
+        public Guid? PlaceId { get; init; }   
+        public EnvironmentType? Environment { get; init; }
+
+        // If true, only return occurrences that have available capacity
+        public bool OnlyAvailable { get; init; }
+    }
+}
