+++ Backend/Api/Infrastructure/Weather/OpenWeatherClient.cs
@@ -42,7 +42,7 @@ namespace Infrastructure.Weather
                     TimeUtc = x.t,
                     TemperatureC = x.x.temp,
                     WindSpeedMs = x.x.wind_speed,
-                    rainVolumeMm = x.x.rain?.OneHour ?? 0.0,
+                    RainVolumeMm = x.x.rain?.OneHour ?? 0.0,
                     ConditionIconUrl = x.x.weather.FirstOrDefault()?.icon ?? "",
                     ConditionText = x.x.weather.FirstOrDefault()?.description ?? "",
                     Source = "openweather3.0"
